+++ src/main/java/com/github/theredbrain/spellengineextension/mixin/spell_engine/internals/SpellHelperMixin.java
@@ -88,7 +88,7 @@ public abstract class SpellHelperMixin {
         throw new AssertionError();
     }
 
-    @Shadow
+    @Shadow(remap = false)
     private static boolean launchSequenceEligible(int index, int rule) {
         throw new AssertionError();
     }
@@ -113,9 +113,9 @@ public abstract class SpellHelperMixin {
                     return SpellCast.Attempt.failMissingItem(new SpellCast.Attempt.MissingItemInfo(ammoResult.ammo().getItem()));
                 }
             }
-            ServerConfig spellEngineExtensionConfig = SpellEngineExtension.serverConfig;
+            ServerConfig spellEngineExtensionConfig = SpellEngineExtension.SERVER_CONFIG;
 
-            if (spellEngineExtensionConfig.spell_cost_health_allowed && ((DuckSpellCostMixin) spell.cost).spellengineextension$checkHealthCost()) {
+            if (spellEngineExtensionConfig.spell_cost_health_allowed.get() && ((DuckSpellCostMixin) spell.cost).spellengineextension$checkHealthCost()) {
                 float healthCost = ((DuckSpellCostMixin) spell.cost).spellengineextension$getHealthCost();
                 if (((DuckSpellCostMixin) spell.cost).spellengineextension$healthCostMultiplierApplies()) {
                     healthCost = healthCost * ((DuckLivingEntityMixin)player).spellengineextension$getHealthSpellCostMultiplier();
@@ -125,7 +125,7 @@ public abstract class SpellHelperMixin {
                     return SpellCast.Attempt.none();
                 }
             }
-            if (SpellEngineExtension.isManaAttributesLoaded && spellEngineExtensionConfig.spell_cost_mana_allowed && ((DuckSpellCostMixin) spell.cost).spellengineextension$checkManaCost()) {
+            if (SpellEngineExtension.isManaAttributesLoaded && spellEngineExtensionConfig.spell_cost_mana_allowed.get() && ((DuckSpellCostMixin) spell.cost).spellengineextension$checkManaCost()) {
                 float manaCost = ((DuckSpellCostMixin) spell.cost).spellengineextension$getManaCost();
                 if (((DuckSpellCostMixin) spell.cost).spellengineextension$manaCostMultiplierApplies()) {
                     manaCost = manaCost * ((DuckLivingEntityMixin)player).spellengineextension$getManaSpellCostMultiplier();
@@ -136,7 +136,7 @@ public abstract class SpellHelperMixin {
                     return SpellCast.Attempt.none();
                 }
             }
-            if (SpellEngineExtension.isStaminaAttributesLoaded && spellEngineExtensionConfig.spell_cost_stamina_allowed && ((DuckSpellCostMixin) spell.cost).spellengineextension$checkStaminaCost()) {
+            if (SpellEngineExtension.isStaminaAttributesLoaded && spellEngineExtensionConfig.spell_cost_stamina_allowed.get() && ((DuckSpellCostMixin) spell.cost).spellengineextension$checkStaminaCost()) {
                 float staminaCost = ((DuckSpellCostMixin) spell.cost).spellengineextension$getStaminaCost();
                 if (((DuckSpellCostMixin) spell.cost).spellengineextension$staminaCostMultiplierApplies()) {
                     staminaCost = staminaCost * ((DuckLivingEntityMixin)player).spellengineextension$getStaminaSpellCostMultiplier();
@@ -147,7 +147,7 @@ public abstract class SpellHelperMixin {
                     return SpellCast.Attempt.none();
                 }
             }
-            if (spellEngineExtensionConfig.spell_cost_effects_allowed && spell.cost.effect_id != null) {
+            if (spellEngineExtensionConfig.spell_cost_effects_allowed.get() && spell.cost.effect_id != null) {
                 StatusEffect effect = (StatusEffect) Registries.STATUS_EFFECT.get(new Identifier(spell.cost.effect_id));
                 if (effect != null) {
                     if (!player.hasStatusEffect(effect)) {
@@ -274,10 +274,10 @@ public abstract class SpellHelperMixin {
                             SpellHelper.imposeCooldown(player, spellId, spell, progress);
                             player.addExhaustion(spell.cost.exhaust * SpellEngineMod.config.spell_cost_exhaust_multiplier);
 
-                            var spellEngineExtensionConfig = SpellEngineExtension.serverConfig;
+                            var spellEngineExtensionConfig = SpellEngineExtension.SERVER_CONFIG;
 
                             // health cost
-                            if (spellEngineExtensionConfig.spell_cost_health_allowed) {
+                            if (spellEngineExtensionConfig.spell_cost_health_allowed.get()) {
                                 float healthCost = ((DuckSpellCostMixin) spell.cost).spellengineextension$getHealthCost();
                                 if (((DuckSpellCostMixin) spell.cost).spellengineextension$healthCostMultiplierApplies()) {
                                     healthCost = healthCost * ((DuckLivingEntityMixin)player).spellengineextension$getHealthSpellCostMultiplier();
@@ -288,7 +288,7 @@ public abstract class SpellHelperMixin {
                             }
 
                             // mana cost
-                            if (SpellEngineExtension.isManaAttributesLoaded && spellEngineExtensionConfig.spell_cost_mana_allowed) {
+                            if (SpellEngineExtension.isManaAttributesLoaded && spellEngineExtensionConfig.spell_cost_mana_allowed.get()) {
                                 float manaCost = ((DuckSpellCostMixin) spell.cost).spellengineextension$getManaCost();
                                 if (((DuckSpellCostMixin) spell.cost).spellengineextension$manaCostMultiplierApplies()) {
                                     manaCost = manaCost * ((DuckLivingEntityMixin)player).spellengineextension$getManaSpellCostMultiplier();
@@ -299,7 +299,7 @@ public abstract class SpellHelperMixin {
                             }
 
                             // stamina cost
-                            if (SpellEngineExtension.isStaminaAttributesLoaded && spellEngineExtensionConfig.spell_cost_stamina_allowed) {
+                            if (SpellEngineExtension.isStaminaAttributesLoaded && spellEngineExtensionConfig.spell_cost_stamina_allowed.get()) {
                                 float staminaCost = ((DuckSpellCostMixin) spell.cost).spellengineextension$getStaminaCost();
                                 if (((DuckSpellCostMixin) spell.cost).spellengineextension$staminaCostMultiplierApplies()) {
                                     staminaCost = staminaCost * ((DuckLivingEntityMixin)player).spellengineextension$getStaminaSpellCostMultiplier();
@@ -378,23 +378,23 @@ public abstract class SpellHelperMixin {
             Spell.ProjectileData.Perks mutablePerks = projectileData.perks.copy();
 
             // region modifying mutable perks
-            ServerConfig serverConfig = SpellEngineExtension.serverConfig;
-            if (serverConfig.spell_projectile_perk_extra_ricochet_attribute_allowed && ((DuckSpellProjectileDataPerksMixin)mutablePerks).spellengineextension$respectExtraRicochetAttribute()) {
+            ServerConfig serverConfig = SpellEngineExtension.SERVER_CONFIG;
+            if (serverConfig.spell_projectile_perk_extra_ricochet_attribute_allowed.get() && ((DuckSpellProjectileDataPerksMixin)mutablePerks).spellengineextension$respectExtraRicochetAttribute()) {
                 mutablePerks.ricochet += (int) (((DuckLivingEntityMixin)caster).spellengineextension$getExtraRicochet());
             }
-            if (serverConfig.spell_projectile_perk_extra_ricochet_range_attribute_allowed && ((DuckSpellProjectileDataPerksMixin)mutablePerks).spellengineextension$respectExtraRicochetRangeAttribute()) {
+            if (serverConfig.spell_projectile_perk_extra_ricochet_range_attribute_allowed.get() && ((DuckSpellProjectileDataPerksMixin)mutablePerks).spellengineextension$respectExtraRicochetRangeAttribute()) {
                 mutablePerks.ricochet_range += (int) (((DuckLivingEntityMixin)caster).spellengineextension$getExtraRicochetRange());
             }
-            if (serverConfig.spell_projectile_perk_extra_bounce_attribute_allowed && ((DuckSpellProjectileDataPerksMixin)mutablePerks).spellengineextension$respectExtraBounceAttribute()) {
+            if (serverConfig.spell_projectile_perk_extra_bounce_attribute_allowed.get() && ((DuckSpellProjectileDataPerksMixin)mutablePerks).spellengineextension$respectExtraBounceAttribute()) {
                 mutablePerks.bounce += (int) (((DuckLivingEntityMixin)caster).spellengineextension$getExtraBounce());
             }
-            if (serverConfig.spell_projectile_perk_extra_pierce_attribute_allowed && ((DuckSpellProjectileDataPerksMixin)mutablePerks).spellengineextension$respectExtraPierceAttribute()) {
+            if (serverConfig.spell_projectile_perk_extra_pierce_attribute_allowed.get() && ((DuckSpellProjectileDataPerksMixin)mutablePerks).spellengineextension$respectExtraPierceAttribute()) {
                 mutablePerks.pierce += (int) (((DuckLivingEntityMixin)caster).spellengineextension$getExtraPierce());
             }
-            if (serverConfig.spell_projectile_perk_extra_chain_reaction_size_attribute_allowed && ((DuckSpellProjectileDataPerksMixin)mutablePerks).spellengineextension$respectExtraChainReactionSizeAttribute()) {
+            if (serverConfig.spell_projectile_perk_extra_chain_reaction_size_attribute_allowed.get() && ((DuckSpellProjectileDataPerksMixin)mutablePerks).spellengineextension$respectExtraChainReactionSizeAttribute()) {
                 mutablePerks.chain_reaction_size += (int) (((DuckLivingEntityMixin)caster).spellengineextension$getExtraChainReactionSize());
             }
-            if (serverConfig.spell_projectile_perk_extra_chain_reaction_triggers_attribute_allowed && ((DuckSpellProjectileDataPerksMixin)mutablePerks).spellengineextension$respectExtraChainReactionTriggersAttribute()) {
+            if (serverConfig.spell_projectile_perk_extra_chain_reaction_triggers_attribute_allowed.get() && ((DuckSpellProjectileDataPerksMixin)mutablePerks).spellengineextension$respectExtraChainReactionTriggersAttribute()) {
                 mutablePerks.chain_reaction_triggers += (int) (((DuckLivingEntityMixin)caster).spellengineextension$getExtraChainReactionTriggers());
             }
             // endregion modifying mutable perks
@@ -403,13 +403,13 @@ public abstract class SpellHelperMixin {
             Spell.LaunchProperties mutableLaunchProperties = data.launch_properties.copy();
 
             // region modifying mutable launch properties
-            if (serverConfig.spell_launch_properties_extra_launch_count_attribute_allowed && ((DuckSpellLaunchPropertiesMixin)mutableLaunchProperties).spellengineextension$respectExtraLaunchCountAttribute()) {
+            if (serverConfig.spell_launch_properties_extra_launch_count_attribute_allowed.get() && ((DuckSpellLaunchPropertiesMixin)mutableLaunchProperties).spellengineextension$respectExtraLaunchCountAttribute()) {
                 mutableLaunchProperties.extra_launch_count += (int) (((DuckLivingEntityMixin)caster).spellengineextension$getExtraLaunchCount());
             }
-            if (serverConfig.spell_launch_properties_extra_launch_delay_attribute_allowed && ((DuckSpellLaunchPropertiesMixin)mutableLaunchProperties).spellengineextension$respectExtraLaunchDelayAttribute()) {
+            if (serverConfig.spell_launch_properties_extra_launch_delay_attribute_allowed.get() && ((DuckSpellLaunchPropertiesMixin)mutableLaunchProperties).spellengineextension$respectExtraLaunchDelayAttribute()) {
                 mutableLaunchProperties.extra_launch_delay += (int) (((DuckLivingEntityMixin)caster).spellengineextension$getExtraLaunchDelay());
             }
-            if (serverConfig.spell_launch_properties_extra_velocity_attribute_allowed && ((DuckSpellLaunchPropertiesMixin)mutableLaunchProperties).spellengineextension$respectExtraVelocityAttribute()) {
+            if (serverConfig.spell_launch_properties_extra_velocity_attribute_allowed.get() && ((DuckSpellLaunchPropertiesMixin)mutableLaunchProperties).spellengineextension$respectExtraVelocityAttribute()) {
                 mutableLaunchProperties.velocity += (int) (((DuckLivingEntityMixin)caster).spellengineextension$getExtraVelocity());
             }
             // endregion modifying mutable launch properties
@@ -463,14 +463,14 @@ public abstract class SpellHelperMixin {
             Spell.LaunchProperties mutableLaunchProperties = data.launch_properties.copy();
 
             // region modifying mutable launch properties
-                    ServerConfig serverConfig = SpellEngineExtension.serverConfig;
-                    if (serverConfig.spell_launch_properties_extra_launch_count_attribute_allowed && ((DuckSpellLaunchPropertiesMixin)mutableLaunchProperties).spellengineextension$respectExtraLaunchCountAttribute()) {
+                    ServerConfig serverConfig = SpellEngineExtension.SERVER_CONFIG;
+                    if (serverConfig.spell_launch_properties_extra_launch_count_attribute_allowed.get() && ((DuckSpellLaunchPropertiesMixin)mutableLaunchProperties).spellengineextension$respectExtraLaunchCountAttribute()) {
                         mutableLaunchProperties.extra_launch_count += (int) (((DuckLivingEntityMixin)caster).spellengineextension$getExtraLaunchCount());
                     }
-                    if (serverConfig.spell_launch_properties_extra_launch_delay_attribute_allowed && ((DuckSpellLaunchPropertiesMixin)mutableLaunchProperties).spellengineextension$respectExtraLaunchDelayAttribute()) {
+                    if (serverConfig.spell_launch_properties_extra_launch_delay_attribute_allowed.get() && ((DuckSpellLaunchPropertiesMixin)mutableLaunchProperties).spellengineextension$respectExtraLaunchDelayAttribute()) {
                         mutableLaunchProperties.extra_launch_delay += (int) (((DuckLivingEntityMixin)caster).spellengineextension$getExtraLaunchDelay());
                     }
-                    if (serverConfig.spell_launch_properties_extra_velocity_attribute_allowed && ((DuckSpellLaunchPropertiesMixin)mutableLaunchProperties).spellengineextension$respectExtraVelocityAttribute()) {
+                    if (serverConfig.spell_launch_properties_extra_velocity_attribute_allowed.get() && ((DuckSpellLaunchPropertiesMixin)mutableLaunchProperties).spellengineextension$respectExtraVelocityAttribute()) {
                         mutableLaunchProperties.velocity += (int) (((DuckLivingEntityMixin)caster).spellengineextension$getExtraVelocity());
                     }
             // endregion modifying mutable launch properties
@@ -478,22 +478,22 @@ public abstract class SpellHelperMixin {
             Spell.ProjectileData.Perks mutablePerks = projectileData.perks.copy();
 
             // region modifying mutable perks
-                    if (serverConfig.spell_projectile_perk_extra_ricochet_attribute_allowed && ((DuckSpellProjectileDataPerksMixin)mutablePerks).spellengineextension$respectExtraRicochetAttribute()) {
+                    if (serverConfig.spell_projectile_perk_extra_ricochet_attribute_allowed.get() && ((DuckSpellProjectileDataPerksMixin)mutablePerks).spellengineextension$respectExtraRicochetAttribute()) {
                         mutablePerks.ricochet += (int) (((DuckLivingEntityMixin)caster).spellengineextension$getExtraRicochet());
                     }
-                    if (serverConfig.spell_projectile_perk_extra_ricochet_range_attribute_allowed && ((DuckSpellProjectileDataPerksMixin)mutablePerks).spellengineextension$respectExtraRicochetRangeAttribute()) {
+                    if (serverConfig.spell_projectile_perk_extra_ricochet_range_attribute_allowed.get() && ((DuckSpellProjectileDataPerksMixin)mutablePerks).spellengineextension$respectExtraRicochetRangeAttribute()) {
                         mutablePerks.ricochet_range += (int) (((DuckLivingEntityMixin)caster).spellengineextension$getExtraRicochetRange());
                     }
-                    if (serverConfig.spell_projectile_perk_extra_bounce_attribute_allowed && ((DuckSpellProjectileDataPerksMixin)mutablePerks).spellengineextension$respectExtraBounceAttribute()) {
+                    if (serverConfig.spell_projectile_perk_extra_bounce_attribute_allowed.get() && ((DuckSpellProjectileDataPerksMixin)mutablePerks).spellengineextension$respectExtraBounceAttribute()) {
                         mutablePerks.bounce += (int) (((DuckLivingEntityMixin)caster).spellengineextension$getExtraBounce());
                     }
-                    if (serverConfig.spell_projectile_perk_extra_pierce_attribute_allowed && ((DuckSpellProjectileDataPerksMixin)mutablePerks).spellengineextension$respectExtraPierceAttribute()) {
+                    if (serverConfig.spell_projectile_perk_extra_pierce_attribute_allowed.get() && ((DuckSpellProjectileDataPerksMixin)mutablePerks).spellengineextension$respectExtraPierceAttribute()) {
                         mutablePerks.pierce += (int) (((DuckLivingEntityMixin)caster).spellengineextension$getExtraPierce());
                     }
-                    if (serverConfig.spell_projectile_perk_extra_chain_reaction_size_attribute_allowed && ((DuckSpellProjectileDataPerksMixin)mutablePerks).spellengineextension$respectExtraChainReactionSizeAttribute()) {
+                    if (serverConfig.spell_projectile_perk_extra_chain_reaction_size_attribute_allowed.get() && ((DuckSpellProjectileDataPerksMixin)mutablePerks).spellengineextension$respectExtraChainReactionSizeAttribute()) {
                         mutablePerks.chain_reaction_size += (int) (((DuckLivingEntityMixin)caster).spellengineextension$getExtraChainReactionSize());
                     }
-                    if (serverConfig.spell_projectile_perk_extra_chain_reaction_triggers_attribute_allowed && ((DuckSpellProjectileDataPerksMixin)mutablePerks).spellengineextension$respectExtraChainReactionTriggersAttribute()) {
+                    if (serverConfig.spell_projectile_perk_extra_chain_reaction_triggers_attribute_allowed.get() && ((DuckSpellProjectileDataPerksMixin)mutablePerks).spellengineextension$respectExtraChainReactionTriggersAttribute()) {
                         mutablePerks.chain_reaction_triggers += (int) (((DuckLivingEntityMixin)caster).spellengineextension$getExtraChainReactionTriggers());
                     }
             // endregion modifying mutable perks