+++ CHANGELOG.md
@@ -6,7 +6,9 @@
 - added 3 new spell schools ("Generic Melee", "Generic Ranged", "Generic Magic")
 - added "generic.magic_damage" entity attribute
 - added 10 new entity attributes which modify cast spells ("generic.extra_launch_delay", "generic.extra_launch_count", "generic.extra_velocity", "generic.extra_ricochet", "generic.extra_ricochet_range", "generic.extra_bounce", "generic.extra_pierce", "generic.extra_chain_reaction_size", "generic.extra_chain_reaction_triggers")
-  
+- removed dependency on Cloth Config
+- added dependency on Fzzy Config
+
 Details about the new content can be found in the readme.
 
 # 1.1.2